Download Pulumi Cron: allow previous version for homebrew and chocolatey
This cron job periodically checks whether package managers have the
updated version of pulumi to install.  This is a very useful check to
make sure our automation works correctly.

However some of the package managers (homebrew and chocolatey in
particular) take a while to act on pulumi-bot's PRs, and thus the
versions are frequently not up to date when this cron job runs shortly
after a release.  This is just how these package managers work and
nothing to worry about.

We shouldn't get an issue every time this happens.  The important
thing here is that the version gets updated eventually.  For homebrew
and chocolatey we therefore also allow the version to be the previous
version.  This means it will take a bit longer for us to notice if a
pulumi isn't updated in these package managers, but the issues this
script opens will then actually be valuable.

(There's still a risk that multiple releases happen within a short
timespan, and the script fails.  This is the case so rarely however
that it's probably not worth worrying about).

An alternative to this would be to check the release date in the
versions.json file, and make a decision based on the age of the latest
release, but that's more complicated to do in bash, and this seems to
be good enough for now.