Add parameterized `Call` support to the engine (#16813)
Parameterization refers to the ability for a provider to vary its schema
based on a parameter that is passed to a new `Parameterize` call on the
provider interface. The package reference that is returned may then be
used to interact with the bespoke schema/packages within.
Paramterization is key to e.g. dynamically bridging providers. In this
instance we can manage and release a single "bridge" provider that
accepts a parameter defining the upstream provider to bridge, and
returns a reference to a dynamically constructed package whose schema
reflects the upstream as needed.

In this commit, we extend the engine so that `Call` calls can accept a
package reference and thus be parameterized.