finish registering resource outputs before registering more resources
Registering resource outputs is happening async in a Goroutine.  This
means this can happen anytime, even after subsequent register resource
calls.  However subsequent register resource calls might rely on these
outputs being registered already, as they won't resolve in their
dependents otherwise.

A future register resource call doesn't have any idea that the outputs
are still to be registered, and thus it's impossible for the SDKs
awaiting funcions, such as Apply or All to wait for the output to be
registered.

Work around this problem by locking ahead of registering a resouce
output, and using an rlock in registerResource.