Document news ignore changes processing in the provider protocol (#16725)
This change adds comments in the protobuf definitions of the
Engine<->Provider protocol to highlight that the `news` value passed to
the provider may be influenced by the state, specifically the old inputs
stored in the state, in an Update scenario, and pre-processed to
implement ignoreChanges at the engine level. This was not entirely
obvious when analyzing provider behavior, and it helps having this
documented in the proto files.

Note that this is especially confusing since the engine is also passing
ignoreChanges to the provider as if to say "please implement it on your
end". This is indeed something that is used to suppress any residual
diffs, say in Terraform-bridged providers, but what's not obvious is
that the engine performs a bit of processing here already.

The relevant implementation seems to be around
https://github.com/pulumi/pulumi/blame/master/pkg/resource/deploy/step.go#L1407

---------

Co-authored-by: Ian Wahbe <ian@wahbe.com>