Rename `auto_secret` to `serialize_as_secret_always`
This commit renames `auto_secret` to `serialize_as_secret_always` and changes the polarity. `serialize_as_secret_always` defaults to `True`, which is the current behavior. A subclass can set it to `False` to make the serialized provider a secret only if any secret Outputs were captured, otherwise, it will not be a secret.