state_move: include stack name in the passhprase prompt
When moving resources between projects that use the passphrase secret
manager, we automatically prompt for the secret, unless
PULUMI_CONFIG_PASSPHRASE or PULUMI_CONFIG_PASSPHRASE_FILE are set.
Unfortunately that can be a bit confusing when moving resources
between different projects, because the prompt does not specify in any
way which stack the passphrase is for.

Fix this by adding a little bit of output to the prompt, showing the
user the stack name for which they are supposed to enter the
passphrase.