Parse conformance programs to decide which packages are needed for SDKs
This is in order to support parameterized packages, which show up in the PCL
source code, but aren't a true separate provider we can list as a provider in
the test case.

Instead of just generating SDKs for each provider listed as used in the test,
we parse and bind the PCL program and ask it what packages it needs. For todays
tests, this is fundementally the same set of packages either way. But once we
add parameterized packages with "package" blocks there could be other packages
returned.