state_move: set up a secret manager in the dest stack if it doesn't exist (#16844)
The destination snapshot currently doesn't necessarily have a secrets
manager set up, e.g. if the stack was just initialized with `pulumi
stack init <stack>`.

At the same time, any secrets we're moving need to be re-encrypted with
the secret manager in the destination stack. If the secret manager is
not defined there this means we'll create a panicCrypter, and `pulumi
state move` will panic.

To prevent this, make the destination stack use the source secret
manager if the secret manager is not already defined there.

Fixes https://github.com/pulumi/pulumi/issues/16835