Add a conformance test for invokes (#16867)
This adds a new provider with just a couple of invokes and a single
resource, and two small projects that uses that invoke to fill in a
stack output.

The first just uses one of the invokes to fill stack outputs. 

The second tests that output'y inputs work and that no arguments to the
invoke work.

Currently disabled for Go because it doesn't handle the multi-return
invoke expression in the stack output value context.