Fix missing overloads to config getters
In https://github.com/pulumi/pulumi/pull/16768 we added overloads to the
config getters when the default is not provided at all, but we did not
provide an overload for the original method signatures where default
defaulted to `None`.

In the presence of overloads, the typechecker ignores the actual
implementation signature.

> The non-@overload-decorated definition, meanwhile, will be used at
> runtime but should be ignored by a type checker

https://docs.python.org/3.12/library/typing.html#overload

Fixes https://github.com/pulumi/pulumi/issues/16877