wait for resource output registration to be done in test (#16897)
This test is racy because resource output registration is currently racy
(see also https://github.com/pulumi/pulumi/issues/16896). While this
race is not fixed, we at least want the test to not be racy. To do that,
we can make the test wait until the resource output registration is
done, before going on with the rest of the test.

Fixes https://github.com/pulumi/pulumi/issues/16536