Use publisher and repository fields to default the Go import path (#16899)
Currently we generate an "internal" package in our Go sdks that the
other packages in the sdk need to import. This requires them to know the
base import path (because Go doesn't have relative imports).

We also want to start writing out go.mod files as part of codegen rather
than provider authors manually adding them in the sdk directory above.
This _also_ requires the import path for the module.

Rather than requiring every user to specify the base import path in the
schema "go language" section we improve the default behaviour to use the
provider and repository fields from the schema that users are probably
filling in anyway.