unmarshal pointers correctly if destination type is anychanges
Currently when unmarshalling resource references into a destination
that's of type `any`, we loose the pointer information, and just
marshal it into the base type, because the destination is not a
pointer type.

However for resource references that's not quite correct, as we do
want to unmarshal it to the pointer type, so it can be a
`pulumi.Resource`, and be used as such.  Make that so here.