Split windows acceptance tests into 2 partitions (#16922)
We are seeing the job “tests on windows-latest/current” fail in the
merge queue somewhat regularly. The job shows up as skipped in the GHA
UI, and the summary has a message

> The hosted runner encountered an error while running your job. (Error
Type: Disconnect).

This usually happens because the machine becomes unresponsive and GitHub
Actions throws out the runner.

Split the windows acceptance tests into 2 partitions to distribute the
load on 2 runners.