Use better names for `package add`-generated SDKs
Currently, `gen-sdk` does not include the name of the plugin or package
in the directory it generates code into. Moreover, it _does_ include the
name of the language for which code is being generated. Neither of these
is a great experience when producing SDKs that must be manually linked,
so we work around this in `package add`, moving generated code into an
`sdks/<package-name>` subdirectory.