Add parameterized providers to Go (#16738)
This adds the first pass of parameterized providers to Go. This only
supports resources, need to expand tests to cover invokes and calls
later.

The method of implementation _might_ be seen as controversial, with all
new methods to do the registrations. An alternative would have been to
make PackageRef a `Resource/InvokeOption`, but that's inconsistent with
how we added it in the other languages, and further the engine won't
send it as a ResourceOption to transforms. So I think this
implementation is probably the preferred way, but while doing this we
might want to just do a `RegisterResourceV2` with `struct` args instead
of positional so if even more arguments need adding in the future we
have a non-breaking way to add them.