Use exact IEEE double min and max values in output test (#16943)
Use the _exact_ value of double min and max in the PCL file for the
number conformance test.

This doesn't change anything currently, but if making any changes to
programgen to try and select between big integers and doubles (for
https://github.com/pulumi/pulumi/issues/14242) the exact value matters.

We try to keep things as plain numbers in programgen where possible, and
only make things into "big integers" (what ever that means per language)
when it needs to be a big integer to maintain precision of the value
given.