Use grpc providers in conformance tests (#16952)
We have a load of engine logic in the grpc layer for providers. Skipping
that and going directly to the Provider objects doesn't give an accurate
test of how the engine will work from these tests.

This is _really_ needed when we want to test things like the engine
options to disable various resource properties (e.g. resource
references, or disable provider preview). Currently all that engine
logic lives in the provider client grpc layer (well it would if it was
complete).

This PR doesn't add tests for any of that but splits the test_host into
its own file and handles the grpc wrapping. It's questionable if we care
that much about these old untested flags (like disable outputs), but we
will want to make use of this change for testing any new flags (e.g.
adding integer support and checking that if it's disabled in the engine
the SDK still works correctly).