Add a warning to `pulumi query` that we plan to remove it
```
$ pulumi query
warning:
================================================================================
query was an experimental command that we have opted to discontinue. It is due
to be removed in a future release before the end of this year (2024).
If you have any feedback or concerns, please let us know by commenting on the
issue at https://github.com/pulumi/pulumi/issues/16964.
================================================================================
```