Fixing a few pylint errors in the python test folder (#16962)
These were found by running `python -m pylint ./lib/test
--rcfile=.pylintrc` in the python SDK folder. Currently we don't
consistently run pylint (or mypy for that matter) against the test
folder and it's built up a number of issues before we can turn it on.

This PR fixes some of the smaller less common ones, a baby step towards
getting the tests fully linted and type checked. Each issue is resolved
in a single commit named for the pylint issue it fixed.