Don't call pluginInstallCommand, defer to schemaFromSchemaSource (#16989)
`schemaFromSchemaSource` is capable of taking multiple types of sources
(plugin, plugin@version, path to plugin). If the plugin is specified
(possible with a version), then it is already capable of downloading the
plugin.

We don't need to call `pluginInstallCommand.Run`, and by doing so we
prevent `pulumi package add` from accepting different source kinds.