create unknowns when the provider is not known during construct (#16868)
Currently when a provider is not known when trying to call construct, we
error out. A provider not being known can however legitimately happen
during a preview, for example when setting up an eks cluster, and then
using that kubeconfig to set up a kubernetes provider.

Users can currently work around that by using `--skip-preview`, and only
going through the actual up, where the provider is known. This is
however not a great user experience.

Return a specific error from the provider Construct call, which we can
then use in the resource monitor to fake constructing a resource with
all unknown outputs. This way the preview can still succeed, but any
dependencies will also end up being unknown.

Fixes https://github.com/pulumi/pulumi/issues/16331