Use events to report progress as system messages
When running a Pulumi operation such as a preview or an update, Pulumi
will detect if plugins (e.g. an AWS provider) are missing and download
and install them appropriately. Presently the user experience when this
happens isn't great (see e.g. #14250), making it hard for the user to
see what is happening/what is taking so long when required plugins are
large.

This commit attempts to rectify this by continuing the work in #16094
that tracks download progress using engine events. In doing so, we are
able to render multiple downloads as part of the existing "system
messages" panel in the Pulumi output, and provide a clean view of what
is going on when downloads must occur before program execution.
Moreover, we generalise that PR to handle any engine process, enabling
us to play a similar trick with plugin installations (which can also
take a while).

To preserve existing behaviour, we introduce a new class for these
events which we call "ephemeral", meaning that they are not persisted or
rendered in contexts such as diffs, for instance. Specifically,
ephemeral events are *not* sent to HTTP backends (i.e. the service), so
this commit should not require any changes to the service before merging
and releasing.

Fixes #14250
Closes #16094
Closes #16937