Prevent double close on OOM detection channel (#17043)
While looking at https://github.com/pulumi/pulumi/issues/17040 I noticed
that the condition in the OOM detection was missing parentheses. This
could potentially make us call `close(o.waitChain)` twice, resulting in
a panic. In practice this does not happen because we only ever see at
most one of the two strings we are looking for.

Note that this does *not* fix
https://github.com/pulumi/pulumi/issues/17040