Introduce Sphinx-backed developer documentation (#16953)
This commit introduces an initial Sphinx configuration for writing
repository-wide developer documentation in Sphinx's MyST Markdown. It
includes Sphinx and MyST configuration as well as workflows for local
development. Documentation on writing documentation -- conventions as
well as how to best utilise Sphinx and other tools is included as a
starting point, with the intention being to fill out the remainder of
the repository in subsequent commits (this will include migrating and
deprecating `developer-docs`). Deployment uses the existing "Read the
docs" webhook and will overwrite the current live version, which at this
point is 6 months' stale anyway.