Print panic banners for snapshot integrity errors
Snapshot integrity errors are generally indicative of a serious issue in
the Pulumi engine. When they occur, we'd like users to file issues with
as much detail as possible so that we can track them down and fix them.
This commit adds a new custom error type for snapshot integrity errors
and a case to the top-level error handler that spots these errors and
prints out a banner asking the user to file an issue.

Fixes #16950