remove changelog comment (#17053)
From our discussion in the ideation meeting yesterday it seemed like
nobody actually finds this comment very useful, and it's just noise on
the PR.  Let's get rid of the comment.  Note that the changelog status
check doesn't actually come from this status check, but separately
from pulumi-bot.  We can safely get rid of this, while keeping that
check.