Prepare golangci-lint upgrade
The latest version of golangci-lint (1.60.3) flags a bunch of new issues in our code base. This PR addresses part of them ahead of the upgrade.

A dynamic string passed to printf style functions as first argument, this can lead to bad % interpolations. The fix is typically to use "%s" as first argument and pass the dynamic string as 2nd argument.
Using os.ModePerm in tests instead of more restricted file permissions. The fix is to use 0o600 for files, or 0o700 for directories.