implement the engine bits for dubugging support
We want to introduce dubugging support for Pulumi programs.  This PR
implements the engine changes necessary for that.  Namely, we
pass the information whether we expect a debugger to be started to the
language runtime, and introduce a way for the language runtime plugins
to report to the engine that we're waiting for a debugger to attach.
The language runtime is expected to include the information relevant
for the user to be able to attach to the debugger, as well as a
shortened message.

The idea is that the configuration can be picked up by an IDE, and the
debugger can attach automatically.  Meanwhile the short message should
contain enough information to be able to attach a debugger manually,
while being short enough to be displayed to the user in the CLI
output. (this will commonly be either the port of the debugger, or the
PID of the process being debugged).

The implementation of the CLI flags and each of the language runtimes
will follow in subsequent PRs.

Co-authored-by: Eron Wright <eron@pulumi.com>