Do not rely on auto api for cloning repos on ds tests (#17071)
Initially I relied `auto.NewLocalWorkspace` to setup the tests workspace
and clone the sample git repo. This is required for testing git specific
logic (like repo url calculation and remote/branch lookup).

The automation api relies on the pulumi binary to to operate but we
don't rely on any of its features on these tests. To simplify it and
remove the need of having the binary in the path, I am replacing it by a
simple logic to clone the repo in a tmp directory.

---------

Co-authored-by: Julien <julien@caffeine.lu>