state_move: allow providers with same URN if inputs match
Currently when there's already a provider in the destination stack
with the same URN as a provider in the source stack (for a resource
that is about to be moved) we error out, unless the IDs match. This is
done so we don't assign a provider to a resource that won't work for
it (e.g. provider with the same name but that have different AWS
region setups).

However we can be a little bit smarter than this, and assign a
resource a provider from the destination stack as long as all the
provider inputs match.  If all the inputs match, we can assume that
the provider will give us a valid situation for the resource.